cleanup python code, fix CI (test)
