+++ .github/workflows/cucumber-tests/action.yml
@@ -1,12 +1,14 @@
 name: "Run cucumber tests"
 description: "Run cucumber tests"
 inputs:
-  tags:
-    description: 'Tags to run'
-    required: false
   feature:
     description: 'Feature file to run'
     required: false
+    default: 'features'
+  tags:
+    description: 'Tags to run'
+    required: false
+    default: '~@flaky'
 runs:
   using: "composite"
   steps:
@@ -14,23 +16,8 @@ runs:
       shell: bash
       run: python3 -m pip install -r src/tests/cucumber_python/requirements.txt
 
-    - name: Run tests on specific tags and specific feature file
-      if : "${{ inputs.tags != '' }}" & "${{ inputs.feature != '' }}"
+    - name: Run tests
       shell: bash
       run: |
         cd src/tests/cucumber_python
         behave --tags ${{ inputs.tags }} ${{ inputs.feature }}
-
-    - name: Run tests on specific tags
-      if : "${{ inputs.tags != '' }}" & "${{ inputs.feature == '' }}"
-      shell: bash
-      run: |
-        cd src/tests/cucumber_python
-        behave --tags ${{ inputs.tags }}
-
-    - name: Run tests on specific feature file
-      if : "${{ inputs.tags == '' }}" & "${{ inputs.feature != '' }}"
-      shell: bash
-      run: |
-        cd src/tests/cucumber_python
-        behave ${{ inputs.feature }}
