+++ src/tests/cucumber/features/steps/study_input_handler.py
@@ -10,25 +10,14 @@ class study_input_handler:
 		self.files_path["general"] 	= self.study_root_dir / "settings" / "generaldata.ini"
 		self.files_path["study"] 	= self.study_root_dir / "study.antares"
 
-	def get_value(self, variable, file_nick_name):
-		# File path
-		file = self.files_path[file_nick_name]
 
-		# Reading the file content (content in)
-		with open(file) as f:
-			# Searching variable and setting its value in a tmp content
-			for line in f:
-				if line.strip().startswith(variable):
-					return line.split('=')[1].strip()
-
-			
 	def set_value(self, variable, value, file_nick_name):
 		# File path
 		file = self.files_path[file_nick_name]
-		
+
 		# Content to print in file (tmp content)
 		content_out = []
-		
+
 		# Reading the file content (content in)
 		with open(file) as f:
 			# Searching variable and setting its value in a tmp content
@@ -37,7 +26,7 @@ class study_input_handler:
 					content_out.append(variable + " = " + value + "\n")
 				else:
 					content_out.append(line)
-		
+
 		# Erasing file content with the tmp content (content out)
 		with open(file, "w") as f:
 			f.writelines(content_out)
